requirements.txt
black==23.9.1
discord.py[voice]==2.3.2
python-dotenv==1.0.0
yt-dlp==2023.7.6
